requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
deepgram-sdk>=3.0.0
deepgram-captions>=1.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
tomli>=2.0.0
aiofiles>=23.0.0
httpx>=0.25.0
numpy>=1.24.0
boto3>=1.28.0
mcp>=1.0.0
